% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialcor.R
\name{calcSpatialCor}
\alias{calcSpatialCor}
\title{Get conditional correlation matrix}
\usage{
calcSpatialCor(
  counts,
  conditionon = NULL,
  neighbors = NULL,
  xy = NULL,
  k = NULL,
  radius = NULL,
  tissue = NULL,
  roundcortozero = 0.1,
  max_cells = 5000,
  verbose = TRUE
)
}
\arguments{
\item{counts}{Single cell expression matrix. Normalizing the data to give every cell the same total expression is preferred.}

\item{conditionon}{Data frame of variables to be conditioned on when computing gene correlation.
Rows correspond to the rows of counts and xy. At a minimum, it is recommended to include cell type
and tissue ID. Including cell signal strength (total counts) and background (negmean) is also recommended.}

\item{xy}{Matrix of xy coordinates}

\item{k}{k for k-nearest neighbor network building}

\item{radius}{Radius for neighbor network building}

\item{tissue}{Used for neighbor network building. Neighbors will only be considered for cell with the same tissue value.}

\item{roundcortozero}{Correlation values with absolute values below this threshold will be stored as 0
to allow for a sparse matrix. If set to NULL, then no rounding to 0 will happen, and a dense matrix will be returned.}

\item{max_cells}{If there are more than this many cells, certain steps will use a random subset of this size.
Output will still be for all cells.}

\item{return_envmat}{logical, for whether to return the environment matrix. Default FALSE.
Note the environment matrix is only computed over the subset.}
}
\value{
A list with 3 elements:
\itemize{
\item condcor: The conditional correlation matrix
\item env: The matrix of neighborhood expression of all genes (columns) in all cells - or a subset of cells (rows)
\item neighbors: a sparse adjacency matrix giving neighbor relationships
}
}
\description{
Builds a neighbor network, calculates an environment matrix, calculates and environment
matrix of variables to be conditioned on, then calculates conditional correlation.
}
