% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineModules.R
\name{calculate_metagene_scores}
\alias{calculate_metagene_scores}
\title{Calculate metagene scores}
\usage{
calculate_metagene_scores(weights, mat)
}
\arguments{
\item{weights}{List of named vectors giving gene weights, such as is output by derive_gene_weights()}

\item{mat}{Expression matrix, cells in rows and genes in columns. Could be either environment or single cell expression.}
}
\description{
Calculate metagene scores
}
