% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTypeAttribution.R
\name{cellTypeAttribution}
\alias{cellTypeAttribution}
\title{score modules for celltype involvement}
\usage{
cellTypeAttribution(
  modulescores,
  weights,
  counts,
  celltype,
  neighbors,
  nsub = 5000,
  verbose = FALSE
)
}
\arguments{
\item{modulescores}{Matrix of each cell's score for each module.}

\item{weights}{List of module gene weights.}

\item{counts}{Single cell expression matrix, cells in rows, genes in columns. The same data used in earlier steps.}

\item{celltype}{Vector of cell types}

\item{neighbors}{Neighbor relationships, stored as a ___________}

\item{nsub}{Subsample size to use. This function only extracts correlations, so 5000 cells is ample}
}
\value{
A list with two elements: "involvescores", a matrix of module * cell type scores, from 0-1.
"attributionmats", a list of matrices giving the involvement of each cell type in each gene in each cluster.
}
\description{
Each module * cell type is given a score, defined by the correlation between environment score
and cell type specific environment gene expression cell type's most
impressive contribution to a gene's expression in module hotspots.
}
