% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetConditionalCovariance.R
\name{cov2cor}
\alias{cov2cor}
\title{convert covariance matrix to correlation:}
\usage{
cov2cor(covmat)
}
\arguments{
\item{covmat}{A covariance matrix}
}
\value{
A correlation matrix
}
\description{
convert covariance matrix to correlation:
}
