% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetConditionalCovariance.R
\name{cova2}
\alias{cova2}
\title{debugged Rfast::cova:}
\usage{
cova2(x, center = FALSE, large = FALSE)
}
\arguments{
\item{x}{matrix}

\item{center}{whether to center}

\item{large}{Same as Rfast::cova large argument}
}
\description{
debugged Rfast::cova:
}
