% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineModules.R
\name{define_gene_weights}
\alias{define_gene_weights}
\title{Define weights for cluster metagenes}
\usage{
define_gene_weights(
  modules,
  neighborhood_expression,
  weighting = "inverse_sqrt"
)
}
\arguments{
\item{modules}{A list, with each element a vector of gene names}

\item{neighborhood_expression}{The neighborhood expression matrix. Colnames must contain the genes given in clusts}

\item{weighting}{One of "inverse_sqrt", "inverse", or "identity"}
}
\value{
A list, where each element is a named vector giving the gene weights for a cluster.
}
\description{
Given a cluster's gene IDs and the environment expression matrix, define how to weight genes for cluster metagene scores
}
