% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineModules.R
\name{defineModules}
\alias{defineModules}
\title{Given a conditional correlation matrix, derive modules, including gene weights}
\usage{
defineModules(
  condcor,
  env,
  min_module_size = 3,
  max_module_size = 20,
  resolution = 0.02,
  corthresh = 0.1,
  min_module_cor = 0.1,
  gene_weighting_rule = "inverse_sqrt"
)
}
\arguments{
\item{condcor}{Conditional correlation matrix}

\item{env}{Environment matrix, possibly for a large subset of cells}

\item{min_module_size}{Modules smaller than this are discarded}

\item{max_module_size}{Modules bigger than than this are subclustered}

\item{gene_weighting_rule}{How to define modules' gene weights, absed on gene expression levels.
One of "inverse_sqrt", "inverse", or "identity".}
}
\value{
A data frame giving module name, gene name, and gene weight, for all genes included in a module.
}
\description{
Given a conditional correlation matrix, derive modules, including gene weights
}
