% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetConditionalCovariance.R
\name{get_conditional_correlation}
\alias{get_conditional_correlation}
\title{get conditional covariance of one matrix given another:}
\usage{
get_conditional_correlation(mat = NULL, condmat = NULL, outputtype = "cor")
}
\arguments{
\item{mat}{Matrix for which conditional covariance is desired}

\item{condmat}{Matrix to be conditioned on}

\item{outputtype}{Either "cor" or "cov"}
}
\value{
A conditional covariance (or correlation) matrix
}
\description{
get conditional covariance of one matrix given another:
}
