% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCorrelationNetwork.R
\name{get_coregulation_network}
\alias{get_coregulation_network}
\title{Get coregulation network}
\usage{
get_coregulation_network(cormat, corthresh = 0.3)
}
\arguments{
\item{cormat}{The correlation matrix}

\item{corthresh}{A scalar. Correlations below this value will not produce graph edges.}
}
\value{
An igraph object
}
\description{
Derive a network graph from a correlation matrix
}
