% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTypeAttribution.R
\name{get_single_module_celltype_involvement}
\alias{get_single_module_celltype_involvement}
\title{score correlation between environment score and per-cell type expression in neighborhood}
\usage{
get_single_module_celltype_involvement(
  scorevector,
  genes,
  counts,
  celltype,
  neighbors
)
}
\arguments{
\item{scorevector}{Vector giving each cell's environment score for the module. Possibly a subset.}

\item{genes}{Vector of module gene names}

\item{counts}{Counts matrix, cells in row, columns holding the genes given in the "genes" argument. Not subsetted.}

\item{celltype}{Vector of cell type assignments. Not subsetted.}

\item{neighbors}{Neighbor relationships, stored as a sparse matrix. Possibly rows are just a subset, but columns are present for all cells.}
}
\description{
For all cell types x genes in the module, reports the correlation between module score and environment expression of the
gene coming from the cell type.
}
