% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineModules.R
\name{name_module}
\alias{name_module}
\title{name module based on gene PCs}
\usage{
name_module(mat)
}
\arguments{
\item{mat}{The conditional correlation matrix for the selected genes}
}
\value{
a name
}
\description{
name a module based on its top 2 genes
}
