% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCorrelationNetwork.R
\name{plotCorrelationNetwork}
\alias{plotCorrelationNetwork}
\title{Plot the conditional correlation matrix as a network}
\usage{
plotCorrelationNetwork(
  x,
  modules,
  genes = NULL,
  corthresh = 0.2,
  show_gene_names = FALSE,
  vertex_size = NULL,
  ...
)
}
\arguments{
\item{x}{Correlation matrix}

\item{modules}{Either the "modules" data frame output by sparc(), or a vector of module names.}

\item{genes}{If \code{modules} is given as a vector, then specify the vector of genes corresponding to it.}

\item{corthresh}{Connect genes with cor > this value}

\item{show_gene_names}{Logical}

\item{vertex_size}{Argument passed to igraph}

\item{...}{Arguments passes to igraph}
}
\value{
A plot.igraph plot
}
\description{
Make an igraph plot from a (presumably conditional) correlation matrix.
Makes an adjacency graph between genes by thresholding the correlation matrix,
then draws an igraph plot, coloring genes by module membership.#'
}
