% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineModules.R
\name{scoreModules}
\alias{scoreModules}
\title{Score cells for module activity, at single cell level and at neighborhood level}
\usage{
scoreModules(counts, weights, neighbors)
}
\arguments{
\item{counts}{Single cell expression matrix}

\item{weights}{List of module weights, as output by defineModules}

\item{neighbors}{Adjacency matrix of neighbors}
}
\description{
Score cells for module activity, at single cell level and at neighborhood level
}
