% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simdat}
\alias{simdat}
\title{Simulated example CosMx data from a kidney sample}
\format{
A list with the following elements:
\itemize{
\item counts A matrix of raw counts, with cells i rows and genes in columns
\item annot A data frame holding cell metadata
\item xy A matrix of cell xy positions
\item celltype Vector of cell types
}
}
\usage{
simdat
}
\description{
Data is simulated so most genes have no spatial pattern (beyond what cell type explains), while a few are highly spatially correlated.
}
\keyword{datasets}
