% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparc.R
\name{sparc}
\alias{sparc}
\title{Derive and score modules of spatially co-expressed genes}
\usage{
sparc(
  counts,
  conditionon = NULL,
  celltype,
  neighbors = NULL,
  xy = NULL,
  k = NULL,
  radius = NULL,
  tissue = NULL,
  min_module_size = 3,
  max_module_size = 20,
  resolution = 0.02,
  corthresh = 0.1,
  min_module_cor = 0.1,
  gene_weighting_rule = "inverse_sqrt",
  roundcortozero = 0.1,
  max_cells = 5000,
  attribution_subset_size = 5000,
  verbose = TRUE
)
}
\arguments{
\item{counts}{Single cell expression matrix. Normalizing the data to give every cell the same total expression is preferred.}

\item{conditionon}{Data frame of variables to be conditioned on when computing gene correlation.
Rows correspond to the rows of counts and xy. At a minimum, it is recommended to include cell type
and tissue ID. Including cell signal strength (total counts) and background (negmean) is also recommended.}

\item{celltype}{Vector of cell type assignments}

\item{neighbors}{Neighbor relationships, stored as a sparse matrix}

\item{xy}{Matrix of xy coordinates.}

\item{k}{k for k-nearest neighbor network building}

\item{radius}{Radius for neighbor network building}

\item{tissue}{Used for neighbor network building. Neighbors will only be considered for cell with the same tissue value.}

\item{min_module_size}{Modules smaller than this are discarded}

\item{max_module_size}{Modules bigger than than this are subclustered}

\item{resolution}{Argument to igraph::cluster_leiden. Lower values produce bigger clusters.}

\item{corthresh}{Only correlations about this value will go into the adjacency graph fed into leiden clustering}

\item{min_module_cor}{Only keep modules with average cor above this value.}

\item{gene_weighting_rule}{How to define modules' gene weights, absed on gene expression levels.
One of "inverse_sqrt", "inverse", or "identity".}

\item{roundcortozero}{Correlation values below this threshold will be stored as 0 to allow for a sparse matrix}

\item{max_cells}{If there are more than this many cells, certain steps will use a random subset of this size.
Output will still be for all cells.}

\item{attribution_subset_size}{Subsample size to use in attribution analysis. This function only extracts correlations, so 5000 cells is ample.}

\item{verbose}{Whether to print progress}
}
\value{
A list, with the following elements:
\itemize{
\item condcor: A sparse matrix holding genes' conditional correlations.
\item modules: A data frame detailing gene module membership and weights
\item scores_env: A matrix of cell * module environment scores
\item scores_sc: A matrix of cell * module single cell scores
\item attributionmats: A list of matrices holding attribution scores for each cell type * gene in each module.
\item celltypeinvolvement: A matrix giving the maximum attribution score for each cell type in each module.
}
}
\description{
Runs the complete workflow for deriving and scoring spatially co-expressed gene modules.
Namely: 1. computes the matrix of each cell's gene expression "environment",
2. derives a matrix of genes' conditional correlation in the environment matrix, conditional
on factors like cell type and tissue ID,
3. defines modules of co-expressed genes from this conditional correlation matrix,
and scores each cell for each module, both in its single cell profile and in its environment.
4. For each module, scores each cell type for its role in driving the contribution to each gene.
}
